seperate redux ui impl into FrontendCore project
