+++ Il2CppInspector.Redux.FrontendCore/Outputs/DisassemblerMetadataOutput.cs
@@ -3,7 +3,7 @@ using Il2CppInspector.Cpp.UnityHeaders;
 using Il2CppInspector.Model;
 using Il2CppInspector.Outputs;
 
-namespace Il2CppInspector.Redux.GUI.Outputs;
+namespace Il2CppInspector.Redux.FrontendCore.Outputs;
 
 public class DisassemblerMetadataOutput : IOutputFormatProvider
 {
