+++ Il2CppInspector.Redux.GUI/Program.cs
@@ -1,3 +1,4 @@
+using Il2CppInspector.Redux.FrontendCore;
 using Il2CppInspector.Redux.GUI;
 using Microsoft.AspNetCore.SignalR;
 
@@ -5,40 +6,23 @@ var builder = WebApplication.CreateSlimBuilder(args);
 
 builder.Services.ConfigureHttpJsonOptions(options =>
 {
-    options.SerializerOptions.TypeInfoResolverChain.Insert(0, AppJsonSerializerContext.Default);
-});
-
-builder.Services.AddSignalR(config =>
-{
-#if DEBUG
-    config.EnableDetailedErrors = true;
-#endif
+    options.SerializerOptions.TypeInfoResolverChain.Insert(0, FrontendCoreJsonSerializerContext.Default);
 });
 
 builder.Services.Configure<JsonHubProtocolOptions>(options =>
 {
-    options.PayloadSerializerOptions.TypeInfoResolverChain.Insert(0, AppJsonSerializerContext.Default);
+    options.PayloadSerializerOptions.TypeInfoResolverChain.Insert(0, FrontendCoreJsonSerializerContext.Default);
 });
 
-builder.Services.AddCors(options =>
-{
-    options.AddDefaultPolicy(policy =>
-    {
-        policy.SetIsOriginAllowed(origin => origin.StartsWith("http://localhost") || origin.StartsWith("http://tauri.localhost"))
-            .AllowAnyHeader()
-            .WithMethods("GET", "POST")
-            .AllowCredentials();
-    });
-});
-
+builder.Services.AddFrontendCore();
 builder.Services.AddSingleton<UiProcessService>();
-builder.Services.AddSingleton<IHostedService>(p => p.GetRequiredService<UiProcessService>());
+builder.Services.AddHostedService(p => p.GetRequiredService<UiProcessService>());
 
 var app = builder.Build();
 
 app.UseCors();
 
-app.MapHub<Il2CppHub>("/il2cpp");
+app.MapFrontendCore();
 
 await app.StartAsync();
